Added guard against duplicated configs Moved EnsureAddonConfig to use SSA
